core: Complete MovieClip error state & Change uncompressed_len to i32
Previously, the MovieClip error state contained one TODO in order to be
completely implemented: The HeaderExt variable uncompressed_len needs to
be -1 in the error state, but its type has been u32.
The type has now been changed to i32, and the variable is set to -1 in
the default_error_header function. Therefore, the MovieClip error state
is now fully implemented.

Other connected functions and variables like SwfMovie::uncompressed_len
or MovieClip::total_bytes have been adjusted to this type change. Code
has been adjusted if needed where this value is used.

+++ core/src/avm2/object/loaderinfo_object.rs
@@ -235,7 +235,7 @@ impl<'gc> LoaderInfoObject<'gc> {
             let should_complete = match self.0.read().loaded_stream {
                 Some(LoaderStream::Swf(_, root)) => root
                     .as_movie_clip()
-                    .map(|mc| mc.loaded_bytes() >= mc.total_bytes())
+                    .map(|mc| mc.loaded_bytes() as i32 >= mc.total_bytes())
                     .unwrap_or(true),
                 _ => false,
             };
